{"schema":"decodex/commit/1","summary":"Modularize settings dashboard","authority":"manual"}
